requirements.txt
streamlit
langchain
langchain-community
langchain-core
langchain-text-splitters
langchain-classic
langchain-groq
langchain-chroma
langchain-huggingface
chromadb
sentence-transformers
pypdf
python-dotenv
